rust/bitbox02/ui: move types to a new file
Keep only functions in ui.rs, so it is easier to do a module switch
for testing after.
